+++ src/main/java/mx/sadead/spring/joinfaces/model/UsuarioRol.java
@@ -26,7 +26,7 @@ public class UsuarioRol implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @Basic(optional = false)
     @Column(name = "rol_id")
     private Long rolId;
@@ -43,15 +43,15 @@ public class UsuarioRol implements Serializable {
     public UsuarioRol() {
     }
 
-    public UsuarioRol(Integer id) {
+    public UsuarioRol(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
